requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arxiv==2.2.0
async-timeout==4.0.3
certifi==2025.4.26
charset-normalizer==3.4.1
distro==1.9.0
exceptiongroup==1.2.2
feedparser==6.0.11
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-core==0.3.56
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langgraph==0.3.34
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.63
langsmith==0.3.37
openai==1.76.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sgmllib3k==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
xxhash==3.5.0
zstandard==0.23.0
